requirements.txt
fastapi
torch
transformers
pillow
numpy
uvicorn
python-multipart